Feat/#85 일정 삭제 (#89)
* [feat - #85] 일정을 찾을 수 없을 때 반환하는 예외 작성

* [feat - #85] id로 schedule을 조회하는 메서드 작성

* [feat - #85] 일정 삭제 서비스 작성

* [feat - #85] 일정 삭제 컨트롤러 작성

* [feat - #85] 일정 삭제 컨트롤러 테스트 작성

* [feat - #85] 일정 삭제 서비스 테스트 작성

* [feat - #85] 일정 패턴의 일정이 더 이상 존재하지 않을 경우 일정 패턴 삭제

* [test - #85] 모든 일정 삭제됐을 경우 일정 패턴도 삭제되는 테스트 작성

* [feat - #85] 일정 단일 삭제와 연관된 일정 전체 삭제 하나의 api로 통합

* [test - #85] api 변경으로 인한 테스트 추가

* [test - #85] ScheduleReadService 테스트 추가

* [test - #85] api 통합으로 인한 ScheduleServiceTest 추가

* [feat - #85] request parameter 기본 값 설정

* [feat - #85] 일정 전체 삭제 방식 변경

- jdbc template에서 jpql 생성 방식으로 변경

* [remove - #85] 필요없는 테스트 삭제

* [test - #85] @Tag 를 이용한 테스트 수정

+++ src/test/java/com/dateplan/dateplan/service/schedule/ScheduleReadServiceTest.java
@@ -183,9 +183,11 @@ public class ScheduleReadServiceTest extends ServiceTestSupport {
 		}
 
 		@AfterEach
-		void tearDown() {
-			scheduleRepository.deleteAllInBatch();
-			schedulePatternRepository.deleteAllInBatch();
+		void tearDown(TestInfo testInfo) {
+			if (testInfo.getTags().contains(NEED_SCHEDULES)) {
+				scheduleRepository.deleteAllInBatch();
+				schedulePatternRepository.deleteAllInBatch();
+			}
 			coupleRepository.deleteAllInBatch();
 			memberRepository.deleteAllInBatch();
 		}
